# 8353638: C2: deoptimization and re-execution cycle with StringBuilder