--- a/src/hotspot/share/opto/runtime.cpp
+++ b/src/hotspot/share/opto/runtime.cpp
@@ -2035,7 +2035,7 @@ static const TypeFunc* make_rethrow_Type() {
 void OptoRuntime::deoptimize_caller_frame(JavaThread *thread, bool doit) {
   // Deoptimize the caller before continuing, as the compiled
   // exception handler table may not be valid.
-  if (!StressCompiledExceptionHandlers && doit) {
+  if (DeoptimizeOnAllocationException && doit) {
     deoptimize_caller_frame(thread);
   }
 }