@dean-long 2025-05-13
This may be a dumb question, but why would we want to optimize throwing NegativeArraySizeException?  Maybe I'm missing something, but I would be surprised if any real apps would be helped by this fix.  I know we have some optimizations for exception throwing, but I would expect those optimizations to be most useful for unavoidable exceptions, like NoClassDefFoundError.