@marc-chevalier 2025-05-13
I agree it's a corner case and it should matter little in real life. Nevertheless, having worst performance in C2 than in C1 is never very nice looking. But to be honest, if the problem were more involved, and the fix more complicated, it's not clear to me whether it would be worth it just for that, indeed.

BUT! This fix should also speedup any exception that happens during allocation. There seem to be quite some `THROW_*` in instanceKlass.cpp, for instance. And this means also for `NoClassDefFoundError`!

`OptoRuntime::new_instance_C` ends with `deoptimize_caller_frame` so would benefit and we have the call chain: `OptoRuntime::new_instance_C` -> `InstanceKlass::initialize` -> `InstanceKlass::initialize_impl` which contains
```cpp
THROW_MSG(vmSymbols::java_lang_NoClassDefFoundError(), ss.as_string());
```