@dean-long 2025-05-13
This change looks good, but do we still need StressCompiledExceptionHandlers after this?  The only other use is in JavaThread::handle_async_exception, and it looks like it should be change too because it's dealing with the same issue.