@marc-chevalier 2025-05-13
I've found a failing test (maybe even two) related to asynchronous exceptions when using this flag. I suspect something is wrong there ([JDK-8356648](https://bugs.openjdk.org/browse/JDK-8356648)). It's also not clear to me whether the failure is directly linked to faulty exceptions handlers: asynchronous exceptions have a complicated (to me) handshake mechanism and maybe it's faulty without the deopts (or deopts makes it less likely). Unlike in [JDK-8004741](https://bugs.openjdk.org/browse/JDK-8004741) (that added `StressCompiledExceptionHandlers`), I'm not hitting the assert "missing exception handler" (that still exists). Aside from being guarded by the same flag, the relation is not clear to me.

For allocations, I couldn't find any problem, and the logic seems simpler. I think it's fine to still use the one that works.